refactor tea settings list as a new stateful widget
